made `image` allow wider and taller images

+++ src/image.c
@@ -5,8 +5,8 @@
 #include <stdio.h>
 
 typedef struct image {
-    uint16_t width;
-    uint16_t height;
+    size_t width;
+    size_t height;
     uint8_t (**pixels)[3];
 } image;
